+++ CHANGELOG.md
@@ -51,6 +51,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   no longer require an explicit revision argument. This walks back a
   deprecation from `jj 0.26`, as the community feedback was mostly negative.
 
+* The revset function `exactly(x, n)` will now evaluate `x` and error if it does
+  not have exactly `n` elements.
+
 ### Fixed bugs
 
 ## [0.33.0] - 2025-09-03